refactor: move drop down menu to separate component
This makes the navigation bar component less clustered
